redefine ExperimentData to include isDummy information for cages
ExperimentData was a map from CageId to CageData.
Now it is a map from CageId to Cage.

WARNING: this will break most things relying on the old form of this data
structure. You won't be able to build unless you comment out sections relying on
the previous form or you fix them.